+++ src/sampletones_application/utils/gui/keyboard/__init__.py
@@ -4,6 +4,7 @@ from sampletones_application.utils.gui.keyboard.router import (
     PRIORITY_MODAL,
     PRIORITY_PANEL,
     PRIORITY_SHORTCUT,
+    ActivePredicate,
     KeyRouter,
     ModalKeyHandler,
 )
@@ -12,6 +13,7 @@ __all__ = [
     "PRIORITY_MODAL",
     "PRIORITY_PANEL",
     "PRIORITY_SHORTCUT",
+    "ActivePredicate",
     "KeyCombination",
     "KeyEvent",
     "KeyRouter",